Dry run
[#157414864]

+++ app/lib/court_reminders_importer.rb
@@ -1,9 +1,9 @@
 module CourtRemindersImporter
-  def self.generate_reminders(court_dates, court_locations)
+  def self.generate_reminders(court_dates, court_locations, options = {})
     time_zone_offset = '-0600' # Make sure this is by instance eventually
 
     Message.transaction do
-      Message.scheduled.auto_court_reminders.destroy_all
+      Message.scheduled.auto_court_reminders.destroy_all unless options[:dry_run]
       court_dates.each do |court_date|
         matching_rrs = ReportingRelationship.where(notes: court_date['ofndr_num'], active: true)
         Rails.logger.info { "[importing court reminders] Found #{matching_rrs.count} RRs with IDs #{matching_rrs.pluck(:id)} with the same ctrack" }
@@ -11,31 +11,37 @@ module CourtRemindersImporter
         rr = matching_rrs.left_joins(:messages).order('messages.send_at DESC NULLS LAST').first
         next if rr.nil?
 
-        court_date_at = Time.strptime("#{court_date['crt_dt']} #{court_date['crt_tm']} #{time_zone_offset}", '%m/%d/%Y %H:%M %z')
-        next if court_date_at < Time.zone.now
+        Rails.logger.tagged("Reminder for Client: #{rr.client.id}") do
+          court_date_at = Time.strptime("#{court_date['crt_dt']} #{court_date['crt_tm']} #{time_zone_offset}", '%m/%d/%Y %H:%M %z')
+          next if court_date_at < Time.zone.now
 
-        body = I18n.t(
-          'message.auto_court_reminder',
-          location: court_locations[court_date['(expression)']],
-          date: court_date_at.strftime('%-m/%-d/%Y'),
-          time: court_date_at.strftime('%-l:%M%P'),
-          room: court_date['crt_rm']
-        )
+          body = I18n.t(
+            'message.auto_court_reminder',
+            location: court_locations[court_date['(expression)']],
+            date: court_date_at.strftime('%-m/%-d/%Y'),
+            time: court_date_at.strftime('%-l:%M%P'),
+            room: court_date['crt_rm']
+          )
 
-        send_at = court_date_at.utc - 1.day
+          Rails.logger.info { "Body: \"#{body}\"" }
 
-        message = Message.new(
-          body: body,
-          reporting_relationship: rr,
-          number_from: rr.user.department.phone_number,
-          number_to: rr.client.phone_number,
-          send_at: send_at,
-          read: true,
-          marker_type: Message::AUTO_COURT_REMINDER
-        )
+          send_at = court_date_at.utc - 1.day
 
-        message.save!
-        message.send_message
+          message = Message.new(
+            body: body,
+            reporting_relationship: rr,
+            number_from: rr.user.department.phone_number,
+            number_to: rr.client.phone_number,
+            send_at: send_at,
+            read: true,
+            marker_type: Message::AUTO_COURT_REMINDER
+          )
+
+          unless options[:dry_run]
+            message.save!
+            message.send_message
+          end
+        end
       end
     end
   end
